+++ models/mappers/book.go
@@ -5,7 +5,7 @@ import (
 	amqp "github.com/kaellybot/kaelly-amqp"
 )
 
-func MapBookJobSetRequest(userId, jobId, serverId string, level uint, lg discordgo.Locale) *amqp.RabbitMQMessage {
+func MapBookJobSetRequest(userId, jobId, serverId string, level int64, lg discordgo.Locale) *amqp.RabbitMQMessage {
 	// TODO
 	return &amqp.RabbitMQMessage{}
 }
